refactor: move project objects to a new folder (objects)
