+++ tasks/gabrielsebrao/ConversaoMoedas/app/src/main/java/com/example/conversaomoedas/objects/Currency.kt
@@ -1,4 +1,4 @@
-package com.example.conversaomoedas
+package com.example.conversaomoedas.objects
 
 import android.widget.Spinner
